put view in the right place
