+++ corehq/apps/app_manager/urls.py
@@ -22,8 +22,8 @@ from corehq.apps.app_manager.views import (
     edit_app_langs, edit_app_attr, edit_app_ui_translations, get_app_ui_translations, rearrange, odk_qr_code,
     odk_media_qr_code, odk_install, short_url, short_odk_url, save_copy, revert_to_copy, delete_copy, list_apps,
     direct_ccz, download_index, download_file, formdefs, release_manager, get_form_questions,
+    overwrite_module_case_list
 )
-from corehq.apps.app_manager.views.forms import overwrite_module_case_list
 from corehq.apps.hqmedia.urls import application_urls as hqmedia_urls
 from corehq.apps.hqmedia.urls import download_urls as media_download_urls
